Organização: Controllers e Rotas (Web)
- Diversas alterações de nomenclaturas
