+++ app/Http/Controllers/CartController.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use Illuminate\Http\Request;
+
+class CartController extends Controller
+{
+    public function cart()
+    {
+        return view('cart');
+    }
+}
