+++ resources/views/cart.blade.php
@@ -1,0 +1,10 @@
+@extends('layouts.front')
+
+@section('content')
+    <br><br><br><br>
+    Carrinho
+    <table>
+        <tr><td>teste</td></tr>      
+        <tr><td> </td></tr>
+    </table>
+@endsection
