+++ resources/views/home.blade.php
